Resurrect StateFile::StateFileControllerConcern, put current_intake + state_name + current_state_code + current_tax_year in there
